+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
+* 2021-02-16 Optimize NcString. Use a single allocation to create the string.
 * 2021-02-16 Add subsliceFrom(), subsliceInRrange(), splitWithLimit(), stringByReplacingStringInRange() in StringSlice & NcString.
 * 2021-02-15 Add NcObject::isKindOf()